requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
future==0.18.2
idna==3.4
oauthlib==3.2.1
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tweepy==4.10.1
twitter==1.19.6
urllib3==1.26.12
